backend/database/requirements.txt
psycopg2-binary
yfinance
nselib
python-dotenv
pandas